+++ src/bolig_ping/scraper.py
@@ -1,27 +1,22 @@
 """Scraping homes available satisfying the given criteria."""
 
+import json
 import logging
-import re
 
-from selenium.common.exceptions import NoSuchElementException, TimeoutException
-from selenium.webdriver.common.by import By
-from selenium.webdriver.remote.webelement import WebElement
+import requests
 from tqdm.auto import tqdm
 
 from .data_models import Home, SearchQuery
-from .webdriver import Webdriver
 
 logger = logging.getLogger(__package__)
 
 
-def scrape_results(search_query: SearchQuery, headless: bool) -> list[Home] | None:
+def scrape_results(search_query: SearchQuery) -> list[Home] | None:
     """Scrape the results of a home search query.
 
     Args:
         search_query:
             The search query to scrape results for.
-        headless:
-            Whether to run the WebDriver in headless mode.
 
     Returns:
         A list of homes that satisfy the search query, or None if no results were found.
@@ -30,88 +25,42 @@ def scrape_results(search_query: SearchQuery, headless: bool) -> list[Home] | No
         HTTPError:
             If there was an error in the HTTP request.
     """
-    url = search_query.get_url()
-    logger.info(f"Fetching URL {url!r}...")
-    webdriver = Webdriver(headless=headless).load(url=url)
+    logger.info("Fetching results...")
 
-    if "Siden findes ikke" in webdriver.text:
+    # Get the results from the search query
+    url = search_query.get_url()
+    response = requests.get(url=url)
+    response.raise_for_status()
+
+    # Parse the response
+    result_dict = json.loads(response.text)
+    results = result_dict["cases"]
+    if results is None:
         return None
 
-    # Close the cookie banner
-    logger.info("Closing cookie banner...")
-    webdriver.click_element_or_ignore(
-        xpath="//button[@id='didomi-notice-disagree-button']"
-    )
-
     # Get the number of pages
-    logger.info("Getting number of results...")
-    try:
-        num_results_elt = webdriver.find_element(
-            "//h1[contains(concat(' ', normalize-space(@class), ' '), ' text-xl ')]"
-        )
-    except NoSuchElementException:
-        raise ValueError(
-            f"Could not find number of results for search query: {search_query}"
-        )
-    except TimeoutException:
-        raise TimeoutError(
-            "Timed out while trying to fetch number of results for search query: "
-            f"{search_query}."
-        )
+    num_results = result_dict["totalHits"]
+    num_pages = num_results // len(results)
+    if num_results % len(results) != 0:
+        num_pages += 1
 
-    num_results_match = re.search(r"[0-9\.]+", num_results_elt.text)
-    if num_results_match is None:
-        raise ValueError("Could not find number of results.")
-    num_results = int(num_results_match.group().replace(".", ""))
-
-    # Extract the homes from the first page
-    logger.info("Scraping first page...")
-    results = webdriver.find_elements(
-        "//div[@data-testid='case-list-card' and "
-        "contains(concat(' ', normalize-space(@class), ' '), ' shadow-card ')]"
-    )
+    # Get the first page of results
     homes = [get_home_from_result(result=result) for result in results]
 
     # Scrape the remaining pages
-    with tqdm(desc="Scraping homes from boligsiden.dk", total=num_results) as pbar:
-        # Update the progress bar
-        pbar.update(len(homes))
-
-        # Calculate the number of pages
-        num_pages = num_results // 50
-        if num_results % 50 != 0:
-            num_pages += 1
-
-        # Iterate over the remaining pages
-        for _ in range(num_pages - 1):
-            # Go to next page page
-            webdriver.click_element_or_ignore(
-                xpath="//ul[@role='navigation']//a[@role='button' and @rel='next']"
-            )
-
-            # Get the results, where we keep trying in case the page hasn't changed
-            # correctly
-            num_homes = len(homes)
-            new_homes = []
-            num_attempts = 3
-            while len(homes) == num_homes:
-                # Get the results
-                results = webdriver.find_elements(
-                    "//div["
-                    "@data-testid='case-list-card' and contains("
-                    "concat(' ', normalize-space(@class), ' '), ' shadow-card ')"
-                    "]"
-                )
+    if num_pages > 1:
+        with tqdm(desc="Scraping homes from boligsiden.dk", total=num_results) as pbar:
+            pbar.update(len(homes))
+            for page_idx in range(2, num_pages + 1):
+                url = search_query.get_url(page=page_idx)
+                response = requests.get(url=url)
+                response.raise_for_status()
+                result_dict = json.loads(response.text)
+                results = result_dict["cases"]
                 new_homes = [get_home_from_result(result=result) for result in results]
                 homes.extend(new_homes)
                 homes = list(set(homes))
-
-                # Monitor the number of attempts and raise an error if we can't change
-                # the page
-                num_attempts -= 1
-                if num_attempts == 0:
-                    raise ValueError("Could not change page.")
-            pbar.update(len(new_homes))
+                pbar.update(len(new_homes))
 
         # Ensure that the progress bar is at 100% at the end
         pbar.n = pbar.total
@@ -119,7 +68,7 @@ def scrape_results(search_query: SearchQuery, headless: bool) -> list[Home] | No
     return homes
 
 
-def get_home_from_result(result: WebElement) -> Home:
+def get_home_from_result(result: dict) -> Home:
     """Get a home from a result.
 
     Args:
@@ -128,68 +77,34 @@ def get_home_from_result(result: WebElement) -> Home:
 
     Returns:
         The home from the result.
-
-    Raises:
-        ValueError:
-            If the result could not be parsed.
     """
-    # Extract URL
-    candidate_urls = [
-        url.get_attribute("href") or ""
-        for url in result.find_elements(By.XPATH, ".//a")
-        if "viderestilling" in (url.get_attribute("href") or "")
-    ]
-    if len(candidate_urls) == 0:
-        raise ValueError("Could not find URL in result.")
-    url = candidate_urls[0].split("?")[0]
-    if not url.startswith("https"):
-        url = "https://boligsiden.dk" + url
+    url = f"https://boligsiden.dk/viderestilling/{result['caseID']}"
+    road_name = result["address"]["roadName"]
+    road_number = result["address"].get("houseNumber")
+    floor = result["address"].get("floor")
+    door = result["address"].get("door")
+    post_code = result["address"].get("zipCode")
+    city = result["address"]["cityName"]
 
-    # Extract address
-    try:
-        address = result.find_element(
-            By.XPATH,
-            ".//div["
-            "contains(concat(' ', normalize-space(@class), ' '), ' bg-black ')"
-            "]//div["
-            "contains(concat(' ', normalize-space(@class), ' '), ' font-black ')"
-            "]",
-        ).text.replace("\n", " ")
-    except NoSuchElementException:
-        raise ValueError(f"Could not find address in result: {result.text}")
+    address = road_name
+    if road_number:
+        address += f" {road_number}"
+    if floor:
+        floor = floor.replace("0", "st.")
+        address += f" {floor}"
+    if door:
+        address += f" {door}"
+    if post_code:
+        address += f" {post_code}"
+    if city:
+        address += f" {city}"
 
-    # Extract span values
-    all_span_values = {span.text for span in result.find_elements(By.XPATH, ".//span")}
-    span_regexes = dict(
-        price=r"kr\.$",
-        size=r"[0-9]+ m²",
-        num_rooms=r"[0-9]+ Vær",
-        monthly_fee=r"Ejerudg.*kr\.?/md",
-        year=r"Opført.*[0-9]{4}",
+    return Home(
+        url=url,
+        address=address,
+        price=result.get("priceCash"),
+        num_rooms=result.get("numberOfRooms"),
+        size=result.get("housingArea"),
+        monthly_fee=result.get("monthlyExpense"),
+        year=result.get("yearBuilt"),
     )
-    values = dict()
-    for span_value in all_span_values:
-        for name, regex in span_regexes.items():
-            if re.search(pattern=regex, string=span_value) is not None:
-                values[name] = extract_number(span_value)
-
-    return Home(url=url, address=address, **values)
-
-
-def extract_number(text: str) -> int | None:
-    """Extract the number from a string.
-
-    Args:
-        text:
-            The string containing the number.
-
-    Returns:
-        The number, or None if no number was found.
-    """
-    match = re.search(r"[0-9][0-9\.]*", text)
-    if match is None:
-        return None
-    match_str = match.group().replace(".", "")
-    if match_str == "":
-        return None
-    return int(match_str)