+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+
+
+## [v1.6.0] - 2025-04-24
 ### Added
 - Now allows multiple emails with the `--email` (`-e`) argument.